Properly handle overriding template section language
